fix: bib export when title is empty
When the bib-heading is set to "" (empty), e.g., via
```
(with-eval-after-load 'ox-hugo
 (plist-put org-hugo-citations-plist :bilbiography-section-heading ""))
```
then `#+print_bibliography:` returns an empty publication list.
The issue is that the current implementation in ox-hugo.el skips the
export on an empty `bib-heading`.

This patch replaces the `when` statement with an `if` statement and adds
a corresponding `else` path to format the bib list without a title.